#226 Упразднение MapHelper
