+++ Zilon.Core/Zilon.Core/Tactics/Behaviour/Bots/AgressiveLogicBase.cs
@@ -152,7 +152,7 @@ namespace Zilon.Core.Tactics.Behaviour.Bots
             var act = actCarrier.Acts.First();
 
             var isInDistance = act.CheckDistance(actorNode.CubeCoords, targetNode.CubeCoords);
-            var targetIsOnLine = MapHelper.CheckNodeAvailability(Map, actorNode, targetNode);
+            var targetIsOnLine = Map.CheckNodeAvailability(actorNode, targetNode);
 
             return isInDistance && targetIsOnLine;
         }
