+++ Zilon.Core/Zilon.Core/Tactics/Spatial/MapBase.cs
@@ -12,7 +12,14 @@ namespace Zilon.Core.Tactics.Spatial
     {
         private readonly IDictionary<IMapNode, IList<IPassMapBlocker>> _nodeBlockers;
 
+        /// <summary>
+        /// Регионы карты.
+        /// </summary>
         public IList<MapRegion> Regions { get; }
+
+        /// <summary>
+        /// Список узлов карты.
+        /// </summary>
         public abstract IEnumerable<IMapNode> Nodes { get; }
 
         protected MapBase()
@@ -22,6 +29,19 @@ namespace Zilon.Core.Tactics.Spatial
             _nodeBlockers = new Dictionary<IMapNode, IList<IPassMapBlocker>>();
         }
 
+        /// <summary>
+        /// Проверяет, является ли данная ячейка доступной для текущего актёра.
+        /// </summary>
+        /// <param name="targetNode">Целевая ячейка.</param>
+        /// <param name="actor">Проверяемый актёр.</param>
+        /// <returns>
+        /// true, если указанный узел проходим для актёра. Иначе - false.
+        /// </returns>
+        /// <exception cref="System.ArgumentNullException">
+        /// targetNode
+        /// or
+        /// actor
+        /// </exception>
         public virtual bool IsPositionAvailableFor(IMapNode targetNode, IActor actor)
         {
             if (targetNode == null)
@@ -47,6 +67,16 @@ namespace Zilon.Core.Tactics.Spatial
             return false;
         }
 
+        /// <summary>
+        /// Указывает, что узел карты освобождён одним из блоков.
+        /// </summary>
+        /// <param name="node">Узел, который будет освобождён указанным блоком.</param>
+        /// <param name="blocker">Блокер, который освобождает узел.</param>
+        /// <exception cref="System.InvalidOperationException">
+        /// Попытка освободить узел {node}
+        /// or
+        /// Попытка освободить узел {node}, который не заблокирован блокировщиком {blocker}
+        /// </exception>
         public void ReleaseNode(IMapNode node, IPassMapBlocker blocker)
         {
             if (!_nodeBlockers.TryGetValue(node, out IList<IPassMapBlocker> blockers))
@@ -62,6 +92,11 @@ namespace Zilon.Core.Tactics.Spatial
             blockers.Remove(blocker);
         }
 
+        /// <summary>
+        /// Указывает, что узел карты занят блоком.
+        /// </summary>
+        /// <param name="node">Узел, который будет занят указанным блоком.</param>
+        /// <param name="blocker">Блокер, который занимает узел.</param>
         public void HoldNode(IMapNode node, IPassMapBlocker blocker)
         {
             if (!_nodeBlockers.TryGetValue(node, out IList<IPassMapBlocker> blockers))
@@ -73,14 +108,47 @@ namespace Zilon.Core.Tactics.Spatial
             blockers.Add(blocker);
         }
 
+        /// <summary>
+        /// Возвращает узлы, напрямую соединённые с указанным узлом.
+        /// </summary>
+        /// <param name="node">Опорный узел, относительно которого выбираются соседние узлы.</param>
+        /// <returns>
+        /// Возвращает набор соседних узлов.
+        /// </returns>
         public abstract IEnumerable<IMapNode> GetNext(IMapNode node);
 
+        /// <summary>
+        /// Создаёт ребро между двумя узлами графа карты.
+        /// </summary>
+        /// <param name="node1">Узел графа карты.</param>
+        /// <param name="node2">Узел графа карты.</param>
         public abstract void AddEdge(IMapNode node1, IMapNode node2);
 
+        /// <summary>
+        /// Удаляет ребро между двумя узлами графа карты.
+        /// </summary>
+        /// <param name="node1">Узел графа карты.</param>
+        /// <param name="node2">Узел графа карты.</param>
         public abstract void RemoveEdge(IMapNode node1, IMapNode node2);
 
+        /// <summary>
+        /// Добавляет новый узел графа.
+        /// </summary>
+        /// <param name="node"></param>
         public abstract void AddNode(IMapNode node);
 
+        /// <summary>
+        /// Выполняет поиск пути к указанному узлу.
+        /// </summary>
+        /// <param name="start">Начальный узел поиска пути.</param>
+        /// <param name="end">Целевой узел поиска пути.</param>
+        /// <param name="context">Контекст поиска пути.</param>
+        /// <param name="outputPath">В результате будет содержать набор узлов,
+        /// представляющих путь из указанного узла в целевой.</param>
+        /// <remarks>
+        /// Передача списка для результатов сделана для оптимизации - не нужно каждый раз создавать список
+        /// и выделять под него память в зависимости от найденного пути.
+        /// </remarks>
         public void FindPath(IMapNode start, IMapNode end, PathFindingContext context, List<IMapNode> outputPath)
         {
             var startNode = start;
@@ -97,5 +165,15 @@ namespace Zilon.Core.Tactics.Spatial
                 }
             }
         }
+
+        /// <summary>
+        /// Проверяет, доступен ли целевой узел из стартового узла.
+        /// </summary>
+        /// <param name="currentNode">Стартовый узел.</param>
+        /// <param name="targetNode">Целевой проверяемый узел.</param>
+        /// <returns>
+        /// Возвращает true, если узел доступен. Иначе, false.
+        /// </returns>
+        public abstract bool CheckNodeAvailability(IMapNode currentNode, IMapNode targetNode);
     }
 }
