+++ Zilon.Core/Zilon.Core/Zilon.Core.csproj
@@ -348,7 +348,6 @@
     <Compile Include="Persons\ITacticalAct.cs" />
     <Compile Include="Tactics\ISectorEntityManager.cs" />
     <Compile Include="Tactics\ITacticalActUsageService.cs" />
-    <Compile Include="Tactics\MapHelper.cs" />
     <Compile Include="Tactics\OpenContainerEventArgs.cs" />
     <Compile Include="Tactics\FixedPropChest.cs" />
     <Compile Include="Tactics\PropContainerManager.cs" />
